+++ docs/experimental/index.rst
@@ -18,3 +18,4 @@ Globus SDK Experimental Components
 
     auth_requirements_errors
     scope_parser
+    consents
