+++ docs/services/auth.rst
@@ -52,15 +52,6 @@ batching, and other functionality.
    :exclude-members: __dict__,__weakref__
    :show-inheritance:
 
-.. autoclass:: ConsentForest
-   :members:
-
-.. autoclass:: ConsentTree
-   :members:
-
-.. autoclass:: Consent
-   :members:
-
 .. autoclass:: DependentScopeSpec
 
 Auth Responses
@@ -86,10 +77,6 @@ Errors
 ------
 
 .. autoexception:: AuthAPIError
-
-.. autoexception:: ConsentParseError
-
-.. autoexception:: ConsentTreeConstructionError
 
 OAuth2 Flow Managers
 --------------------
