+++ src/globus_sdk/experimental/consents/__init__.py
@@ -1,0 +1,10 @@
+from ._errors import ConsentParseError, ConsentTreeConstructionError
+from ._model import Consent, ConsentForest, ConsentTree
+
+__all__ = [
+    "Consent",
+    "ConsentTree",
+    "ConsentForest",
+    "ConsentParseError",
+    "ConsentTreeConstructionError",
+]
