+++ src/globus_sdk/services/auth/errors.py
@@ -1,12 +1,8 @@
 from __future__ import annotations
 
 import logging
-import typing as t
 
 from globus_sdk import exc
-
-if t.TYPE_CHECKING:
-    from globus_sdk.services.auth.consents import Consent
 
 log = logging.getLogger(__name__)
 
@@ -42,19 +38,3 @@ class AuthAPIError(exc.GlobusAPIError):
                 self.request_id = maybe_error_id
 
         return True
-
-
-class ConsentParseError(Exception):
-    """An error raised if consent parsing/loading fails."""
-
-    def __init__(self, message: str, raw_consent: dict[str, t.Any]):
-        super().__init__(message)
-        self.raw_consent = raw_consent
-
-
-class ConsentTreeConstructionError(Exception):
-    """An error raised if consent tree construction fails."""
-
-    def __init__(self, message: str, consents: list[Consent]):
-        super().__init__(message)
-        self.consents = consents
